♻️ extract PlaybackManager interface; rename existing class to PlaybackManagerImpl
