+++ composeApp/src/desktopMain/kotlin/com/calypsan/listenup/client/di/PlatformModule.kt
@@ -15,6 +15,7 @@ import com.calypsan.listenup.client.playback.AudioTokenProvider
 import com.calypsan.listenup.client.playback.DesktopPlaybackController
 import com.calypsan.listenup.client.playback.PlaybackController
 import com.calypsan.listenup.client.playback.PlaybackManager
+import com.calypsan.listenup.client.playback.PlaybackManagerImpl
 import com.calypsan.listenup.client.playback.ProgressTracker
 import com.calypsan.listenup.client.playback.SleepTimerManager
 import com.calypsan.listenup.client.playback.FfmpegAudioPlayer
@@ -120,8 +121,8 @@ val platformModule: Module =
         }
 
         // Playback manager (with codec negotiation enabled)
-        single {
-            PlaybackManager(
+        single<PlaybackManager> {
+            PlaybackManagerImpl(
                 serverConfig = get(),
                 playbackPreferences = get(),
                 bookDao = get(),
