+++ shared/src/macosMain/kotlin/com/calypsan/listenup/client/di/PlaybackModule.macos.kt
@@ -13,6 +13,7 @@ import com.calypsan.listenup.client.playback.AudioPlayer
 import com.calypsan.listenup.client.playback.AvFoundationAudioPlayer
 import com.calypsan.listenup.client.playback.PlaybackController
 import com.calypsan.listenup.client.playback.PlaybackManager
+import com.calypsan.listenup.client.playback.PlaybackManagerImpl
 import com.calypsan.listenup.client.playback.ProgressTracker
 import com.calypsan.listenup.client.playback.SleepTimerManager
 import com.calypsan.listenup.client.sync.BackgroundSyncScheduler
@@ -97,8 +98,8 @@ val macosPlaybackModule: Module =
         }
 
         // Playback manager
-        single {
-            PlaybackManager(
+        single<PlaybackManager> {
+            PlaybackManagerImpl(
                 serverConfig = get(),
                 playbackPreferences = get(),
                 bookDao = get(),
